Update patch set 1
Patch Set 1:

I would suggest submitting this up to AOSP. Since it is just a warning. I do not want to take it as it will just be a merge headache in the potential future. Even though its trivial to resolve, a lot of trivial merges every new release adds up to a lot of work :)

Patch-set: 1